source/requirements.txt
discord.py
flask
python-dotenv
requests
